backend/requirements.txt
livekit-agents[openai,silero,deepgram,cartesia,turn-detector]~=1.5
livekit-api>=1.0.7
chromadb>=0.5.0
openai>=1.30.0
pymupdf>=1.24.0
fastapi>=0.111.0
uvicorn>=0.30.0
httpx>=0.27.0
python-dotenv>=1.0.0